requirements.txt
aiohttp==3.7.4.post0
aiosmtplib==1.1.6
aiounittest==1.3.0
arrow==1.1.1
asgiref==1.1.1
async-timeout==3.0.1
asyncio==3.4.3
bidict==0.21.2
cachetools==4.2.2
chardet==4.0.0
click==8.0.1
cssselect==1.1.0
cssutils==2.3.0
email-to==0.1.0
Flask==2.0.1
Flask-Cors==3.0.10
itsdangerous==2.0.1
lxml==4.6.3
Markdown==3.3.4
multidict==5.1.0
premailer==3.10.0
python-engineio==4.2.0
python-socketio==5.4.0
Werkzeug==2.0.1
yarl==1.6.3
